:recycle: Make handling of RestEnvironments simpler by reduction.
Instead of having a limited subset of `RestEnvironment`s where the user
must choose between one of staging, production, or local, this makes
Bubbles have a single `environment` that's meant to be used within a
configuration block. This enables users to utilize it from within an
environment file initialized during startup of a Rails application.

Refs #4.