:white_check_mark: Update disconnected host test to eject VCR cassette.
We were previously using a VCR cassette that shouldn't exist for the
specification:

  > Bubbles:Resources when using the local environment or a previously
  > recorded call when accessed with a POST request when the host is
  > unavailable should fail gracefully

This is somewhat sub-optimal, because the user may have inadvertently
created the VCR cassette with the given name. Instead, we now use
`VCR.eject_cassette`, which makes it so that VCR _always_ plays a live API
call, rather than re-using a recorded one.

Refs #19.