:bug: Make unauthenticated HEAD requests use the API key from the env.
In previous iterations of the library, the RestEnvironment can be populated
with an API key, allowing this data to be defined _once_, rather than
having to pass it as part of every single call to the API. Unfortunately,
calls with the method of HEAD were never updated to utilize this. This
commit makes that update.

Fixes #19.