:recycle: Move running of RSpec via Rake into its own task file.
We want to separate individual Rake tasks into their own files, for
organization purposes. When we had just one thing defined in our Rakefile,
it was ok that it was defined directly in the Rakefile. Now that we're
going to be adding code coverage tasks, though, we want to enforce better
code organization. This isolates the specification running into its own
.rake file in the lib/tasks/ directory.

Refs #26.