:white_check_mark: Make VCR match against some subset of headers
We had disabled checking against headers when using VCR previously, because
we couldn't scale it. Unfortunately, when refactoring the code related to
header population, this doesn't allow us any certainty that we haven't
broken anything. In order to fix this, we added custom matcher blocks for
VCR in the configuration.

Refs #10.