:sparkles: Add support for unauthenticated POST requests without a key
Unauthenticated POST requests that do not provide an API key were not
previously supported. This is not a common protocol to use, and, in fact,
we were unable to find an open API that supported them. The tests thus use
a contrived formulation for testing this, but it could conceivably be
necessary for an intranet-only API.

Refs #12.
Refs #16.