:sparkles: Add a method to that conditionally returns the api_key.
In some cases, the API is necessary, but this determination is based on
whether or not the Endpoint actually requires it. Since the Endpoint is
created asynchronously from the RestEnvironment, it's possible that the
RestEnvironment might not know at time of creation whether the API key
should be returned. This method corrects that by adding a dependency for
this determination on the Endpoint.

This also verifies that all REST requests now send the API key,
conditionally.

Refs #12.