:sparkles: Add the ability to process unauthenticated DELETE requests
Delete requests without any authentication should be very rare. It seems as
though this shouldn't really be allowed, as it's bad practice (who would
want to allow the deletion of records without any form of authentication?).
That said, it IS a legal REST method, and so, in the vein of supporting all
possible things that a client would want to do, we add this as a potential
method for bubbles to process.

Fixes #16.