:recycle: Extract method for composing headers
This is actually the first of a couple of steps to getting the headers a
little more organized within the codebase. The next step is to take the
header processing out of RestClientResources entirely and put them into a
derived class, so we can get _ALL_ headers for a given Endpoint and
Environment combination.