:sparkles: Add a method for having multiple APIs within the same instance.
Bubblez now supports having multiple APIs simultaneously. Prior to this
commit, it was only possible to have a single API client (i.e. you wouldn't
be able to use a single Bubblez instance/configuration for accessing both
Google and Stripe). Now, however, we've switched to named instances of API
configurations. This allows us to have as many APIs as we want, as long as
each one has a unique name.

Refs #45.